Some refactoring due to auto generated classes
