+++ modules/citrus-jdbc/src/test/java/com/consol/citrus/jdbc/server/JdbcEndpointAdapterControllerTest.java
@@ -24,9 +24,9 @@ import com.consol.citrus.jdbc.data.DataSetCreator;
 import com.consol.citrus.jdbc.message.JdbcMessage;
 import com.consol.citrus.jdbc.message.JdbcMessageHeaders;
 import com.consol.citrus.jdbc.model.JdbcMarshaller;
-import com.consol.citrus.jdbc.model.Operation;
 import com.consol.citrus.message.Message;
 import com.consol.citrus.message.MessageType;
+import com.consol.citrus.model.message.jdbc.Operation;
 import org.springframework.xml.transform.StringResult;
 import org.testng.annotations.BeforeSuite;
 import org.testng.annotations.Test;
